requirements.txt
asgiref==3.6.0
Django==3.2.16
django-cleanup==7.0.0
django_debug_toolbar==3.8.1
django-environ==0.9.0
Pillow==9.4.0
pytz==2022.7.1
sorl-thumbnail==12.9.0
sqlparse==0.4.3